Make eslint fail on warning
All the fixed max-len warnings occur in a subdirectory that has its own
eslint config file which has a `rules` key and thus overrides the
`"max-len": "off"` directive in the root eslint config file.
That might not be intentional, but I think all these changes improve
readability. Adjusting our eslint config inheritance setup can be done
in a different change.

The mocha/no-setup-in-describe was a false positive.

Change-Id: I862a4e66d56d5aebfeb0b342cac65ae09b71e31f