Auto-reconfigure LBFactory in DumpGenerator
Reload the database configuration at the beginning of each batch, so
that we don’t keep connecting to depooled replicas for the duration of
the whole dump. Put this into a configurable callback, so that we don’t
need to inject a DomainDB into DumpGenerator (and pass it through the
subclass constructors); this also makes the testing easier.

We call the callback at the *beginning* of each batch so that it’s
guaranteed to be called at least once even when there is nothing to
dump; this means we can safely expect at least one call in
RdfDumpGeneratorTest even if one of the test cases happens to dump
nothing (I didn’t check). In JsonDumpGeneratorTest, the tests are
arranged differently and I added a separate test method instead; note
that it expects one more call than testProgressReporter(), since the
last, empty batch of item IDs calls the callback but reports no progress
anymore.

Bug: T321770
Change-Id: Ibe02bf7506b5b4f5e406b5d896ad057e345724de
Depends-On: I7ec3a2dde3c60b1ca488f62004f996c8eb0c57ca