REST: Validate PropertyValuePairs during deserialization
The data model's `SnakDeserializer`, that was previously used in the REST API, only throws an error if the serialization is of an incorrect format. There are no checks to ensure the resulting `Snak` is valid (e.g. that the Property exists). This includes creating an `UnDeserializableValue` if the `DataValueDeserializer` throws a `DeserializationException`. The `SnakValidator` is used to validate if a `Snak` (or an object containing a `Snak`) is in fact valid.

When creating the REST API Deserializers, the decision was made for the deserializers to do validation at the same time to ensure an invalid object wasn't returned. The only validation that is in the `SnakValidator` and not in the `PropertyValuePairDeserializer` is value validation via the `DataTypeValidatorFactory`.

This patch adds the missing value validation to the `PropertyValuePairDeserializer` and removes the, no longer needed, `SnakValidator` from `JsonDiffStatementPatcher` and `SnakValidatorStatementValidator`. The latter is also renamed to `StatementDeserializerStatementValidator`.

Task T323779 has been created to decide if the DataAccess directory is still the best place for this validator.

Change-Id: Ib6b72430252a15cb8db40dbaf00be924bbcaa340