Introduce new showcalendar option for time formatters
Previously, the time formatters for non-HTML formats (plain text and
wikitext) never showed the calendar model – Wikibase’s logic to
automatically show the calendar model only if necessary was only
available in the HTML formats.

With this change, showing the calendar model becomes configurable. The
logic determining whether showing the calendar is needed is moved to a
separate class, ShowCalendarModelDecider, and the new formatter option
'showcalendar' can be used to enable this functionality in all time
formatters by setting it to 'auto'; you can also set it to true or false
to always or never show the calendar model.

To preserve compatibility with older Wikibase versions, the default of
the option is 'auto' in HtmlTimeFormatter, but false in
PlaintextTimeFormatter, so there is no change in the default behavior;
users who want the “auto” behavior in the non-HTML formatters need to
opt into it by setting the 'showcalendar' option to 'auto', for example
via the options parameter of the wbformatvalue API. (In the Lua
interface, there’s no way to customize the options, AFAICT.)

The HtmlTimeFormatter tests are copied to ShowCalendarModelDeciderTest,
and most of them are then removed from HtmlTimeFormatterTest since they
don’t need to be tested twice; the integration between the two is also
tested by some test cases in WikibaseValueFormatterBuildersTest.

Bug: T323568
Change-Id: Icb60fe27955b1349d95591381fd9115f8de882cc