Use SelectQueryBuilder in WikiPageEntityStore
And replace the now-deprecated ActorMigration with ActorNormalization.
(This was our only use of ActorMigration in Wikibase, yay.)

Phan complains that the cast of $actorId to int is redundant, which is
true because findActorId() has a static return type (not just phpdoc) of
?int. However, if that return type ever changes, using the actor ID in
the query without casting it to an int (or explicitly quoting it) might
result in an SQL injection vector, so I prefer to be safe here and still
keep the redundant cast; therefore, suppress the Phan warning.

Bug: T312510
Change-Id: I42d56fe549bc64369b9e9b77034507bd5455acfc