REST: Use requested fields in ItemData
This change makes the use of requested fields in ItemData more explicit
by listing them in a required constructor parameter. This way a new
'isRequested( $field )' function can be used to determine if a field
should be included in the ItemDats's serialization, rather than relying
on the implicit information whether it is null.

Change-Id: Idc4774026c0485a70075e63fd777db77439efccf