Delay work in EntityConstraintProvider
Most edits don’t touch properties, so it seems unnecessary to create the
label uniqueness validator in the constructor already – let’s postpone
it until we need it in getUpdateValidators(). (I think this method is
only called once per edit, so I don’t think creating a new instance vs.
reusing it is a concern; if it was a concern, then the more important
instance to save would probably be the SiteLinkUniquenessValidator.)

Also use PHP 7.4 typed properties.

Bug: T320490
Change-Id: I4d16dc0ccad0546c859c4ab8e86cb364df9a4b99