Change sitelink conflict message if redirects supported
If the wiki has redirect badge items configured and thus supports
sitelinks to redirects, use a variation of the error message if a
sitelink conflict is detected: the user may want to add a redirect badge
to their edit. (For the user experience, it would be nicer if the
message depended on whether the sitelink being added had a redirect
resolved or not, but it would require a lot more work to carry this
information from the SiteLinksChangeOpDeserializer all the way to the
SiteLinkUniquenessValidator.)

Also declare strict types, and fix a missing $services argument in the
service wiring.

Bug: T320490
Change-Id: I1b01d02c11d1b8a56eb607eb27609647dd381ce3