REST: Enforce dependency rules with phpat
This patch introduces PHPat [0] to enforce the dependency rules of our
architecture. PHPat is an extension of the static analysis tool PHPStan
[1].

Due to MediaWiki's extensibility, autoloading is more complicated
compared to other PHP applications, and did not work out of the box
using PHPStan. Passing the PHPUnit bootstrap file as a custom autoloader
[2] when invoking `phpstan` made it work, though.

Unfortunately, PHPat recently removed its inclusive `canOnlyDependOn()`
assertion [3] in favor of exclusive `shouldNotDependOn()`. This does
not work well in our case since the core of our application is strictly
limited to a few namespaces. Given rules such as "domain models may
depend on classes in its own namespace and DataModel entities", it's a
lot easier and less error prone to list allowed namespaces, than to
list all the forbidden ones.

This is worked around here by chaining
`shouldNotDependOn()->classes( Selector::all() )->excluding( /*...*/ )`
which forms the "allow list" functionality that we want via this "should
not depend on anything, except..." construct.

The new test only covers a few layers of our architecture but has
proven itself to be useful already as it discovered a dependency issue
concerning a misplaced validator interface
(I832ec7d4894acc5bb55934e77b8a9f081253f3bb). Tests for the validation,
serialization and presentation namespaces are missing, but can be added
in other patches later.

[0]: https://github.com/carlosas/phpat
[1]: https://phpstan.org/
[2]: https://phpstan.org/user-guide/discovering-symbols#custom-autoloader
[3]: https://github.com/carlosas/phpat/issues/192

Bug: T324093
Change-Id: Ib077728c28c95261b0c01cd067572bd6f16ada82