REST: Refactor return in switch statement
PHPStan can't handle the return in the switch statement without a
default case and complained with the following error:

Method PropertyValuePairDeserializer::deserialize() should return
Wikibase\DataModel\Snak\Snak but return statement is missing.

At this point in the code we're guaranteed that the checked value is one
of "novalue", "somevalue" or "value", so we can just throw a
LogicException in the default case.

Change-Id: I454c80a0bb56532b256f684e6ad96789a4080d75