Fix spacing of language fallback indicator
The spacing was originally added (together with the indicator in
general) in change I3779175661 (commit f4a59c5781), as a margin-left on
the element. Soon afterwards, change Ic87b387917 (commit 39e2074292)
changed this to a combined margin-left and margin-right, because with
mixed LTR and RTL text on a page, the indicator may be on either side of
the text that it logically follows. (That is, on an overall RTL page,
where CSSJanus mirrored the margin-left into margin-right, the indicator
may be after a LTR label, which means the indicator is to the right of
the label and needs space on its left. See T89047 for screenshots.) This
double margin, where one of them is redundant (but it can be hard to
know *which one*), is okay in the original context of statement values
and entity titles, but it’s not so great in WikibaseQualityConstraints,
where we sometimes put punctuation directly after the indicator, which
makes the extra margin visible.

Later, in change I52dac7c042 (commit 358c0584c2) and change Id36bb8f0fb
(commit ffcf369ae9), the left part of this margin was changed to a
no-break space, so that double-clicking the main text doesn’t select the
indicator as well. The right part of the margin was kept at the time;
however, this means that the no-break space and the margin can end up on
the same side and stack up, which is undesirable. Also, because the
no-break space was added outside the indicator element, it is visible
even when the indicator isn’t (e.g. because the indicator is only an
indicator for a fallback language, which we don’t show by default); this
is the specific problem we’re trying to solve now (T291608).

This change moves the no-break space inside the element, but also
removes the remaining half of the margin. As far as I can tell, the
no-break space makes the margin obsolete in all cases. The indicator
element (<sup>) has no lang= nor dir= attribute (because we don’t know
in which language the language name is given – it may be in the user
interface language, one of its fallback languages, or an autonym), so
the bidi behavior of the indicator is only determined by the character
types of its contents, and the Unicode Bidirectional Algorithm seems to
do the right thing with the space (that is, put it visually between the
label and the rest of the indicator) in all the combinations I’ve
tested. Note that this changes if the indicator as a whole gets a dir=
attribute, so add a comment explaining that we shouldn’t do that: if we
ever add a dir=, it must be on a separate element inside the indicator.

Bug: T291608
Change-Id: I8dada901bf351b4b5bcf2ead8a3672e24b914abd
Depends-On: I388733b02bcb68a60a97ffa0488dbf857dad271b
Depends-On: Idf1af23753fa8bca2f1d444fecb9261a9f91e2f3