Make DispatchingValueFormatter lazy
Allow the formatters to be factory functions instead of direct
ValueFormatter instances, and create them on demand (saving the created
instances for reuse afterwards). This means that incorrectly wired up
formatters (not ValueFormatter instances) will now only be detected in
format() / formatValue(), not in the constructor anymore.

In OutputFormatSnakFormatterFactoryTest, we now need to call
formatSnak() to ensure the assertion in the formatter callback actually
runs (otherwise the test is “risky” because it performs no assertions).

Bug: T326135
Change-Id: I1ddb4467c0799762b173914b3dfaaea4ab537999