Include "parent" parameter to HTMLFormField for upcoming core change
There is a MW core patch proposes making the 'parent' field of
the parameters to HTMLFormField mandatory (or at least deprecating
not doing it). Change the test classes to provide this parameter
as a mock of the HTMLForm class. See T326456 for more details/rationale.

Note: This makes one code change. The parent form will no longer
override the default language for HTMLContentLanguageField if both
are set. I believe that the previous behaviour was a typo in the code.
There doesn't appear to be any wikibase code outside of tests that
either don't set or parent, or set an explicit default language,
so this should have no impact other than making the test work.

Bug: T326456
Change-Id: I15a39605e3eec8a5c265c4a331039fa906eda036
Needed-By: Ica0740049f0a3e8ec764903c5b71825e4d628a3f