Remove LanguageWithConversion methods returning Language
LanguageWithConversion had two pairs of fields and getters for the
actual and source language: the language code (string), and a Language
instance; the latter had to be obtained using a LanguageFactory.
However, the Language instance getters were only used in tests and in
SpecialMyLanguageFallbackChain; it’s more elegant to remove these
getters, and either use the language code directly or inject the
necessary LanguageFactory in SpecialMyLanguageFallbackChain instead.

Note that, in general, $languageFactory->getLanguage( $code )->getCode()
is *not* equivalent to just $code, since LanguageFactory applies the
mapping from the $wgDummyLanguageCodes. However, we do the same thing in
LanguageWithConversion::factory() via validateLanguageCode(), so for any
language code that we got from a LanguageWithConversion, using it to
create a Language instance and getting its code is certain to be
equivalent to just using the code in the first place.

Change-Id: Ie2961a7a169b3a91e2b290e8b6be9bac306203ce
Depends-On: Id92863731cc268ba519c19587735b6278551d789