REST: Introduce NewStatementReadModel
This builder inherits from the NewStatement builder for DataModel
Statement objects and simply converts them when calling `build()`. If
DataModel Statement and read model Statement become too different, we
might want to undo this coupling of the test builder, but for now it
works well enough.

This patch does not yet make use of the new builder in
`WikibaseEntityLookupItemDataRetrieverTest` since that test still relies
on DataModel Statement objects being added to Item objects, for which we
don't have a read model yet. We can clean that up once T326635 is done.

Change-Id: I5f10fdffe652944a2a082ff1d455c8e4cc403efa