Add API integration test for change dispatching
Based on the repo+client test in the Wikibase release pipeline,
specifically the wikitext version [1] rather than the Lua version [2],
but implemented as an API integration test and so done fully through the
Action API (compare change I501ec8369d for a WebdriverIO-based version).

Most of the work happens in various before() handlers; the tests should
be independent of each other: you can .skip or .only run any of them, if
you want. (Note that .skip and .only also work on describe().)

[1]: https://github.com/wmde/wikibase-release-pipeline/blob/wmde.9/Docker/test/selenium/specs/repo_client/item.js
[2]: https://github.com/wmde/wikibase-release-pipeline/blob/wmde.9/Docker/test/selenium/specs/repo_client/extensions/scribunto-item.js

Bug: T326082
Change-Id: If4a281bc2ec91823deab5893bab6927f9ee75ce8