Remove obsolete check for $wgKartographerEnableMapFrame
We are going to remove this setting via I67e8d27. This code will
then stop working because it checks if the setting exists.

The check was never really needed. The setting defaults to true
and always was true on wikidata.org, as far as I know. It would be
possible to leave compatibility code behind that works with and
without this setting. But I think this is not necessary for the
reasons above.

Bug: T326288
Change-Id: I70ad9795e4f519cd3c2b19e380c4038fb594ef6d