REST: Disallow use of fully qualified class names
This enables a custom MediaWiki sniff that disallows fully qualified
class names outside of the "use …" section. It explicitly includes use
of the main namespace and class names in "extends" and "implements"
statments.

We find it useful to have all dependencies listed in the use statements,
because this way it will be easier to identify architecture constraint
(aka 'onion layer') violations.

Change-Id: Iea9915a07da65ac69547fa3098b353d86809dd98