Link to the Site subgroup in the language selector
The `#sitelinks` heading actually has `display: none;` and so linking to
it does not actually scroll the page. This now links to the current
site's site-group. That has the added advantage that the user is already
at the right place, even if there are a lot of sites in other
site-groups.

Since this functionality should also be available on mobile, we have to
explicitly set that as a target for the SitesModule, this coincides with
the rework of that system in T326405.

Bug: T310259
Change-Id: Icd9bb0496605416a213771a2e64c1daf4632d33c