Let DateFormatParser accept and skip redundant day of the week
The idea is that every possible date format that displays the day of
the week also displays day, month, and year. The day of the week is
not needed to identify the exact date. It can as well be ignored.

This is especially relevant in many Asian contexts like zh-tw, yue,
and many more. It appears like it's very common that these formats
end with a (D) or (l). These can currently not be copy-pasted to
Wikidata, neither with nor without the day of the week, because the
format is entirely rejected by the parser. After this change these
should both be accepted on Wikidata:

* 2014年09月01日
* 2014年09月01日 (一)

Possible alternative implementations:
1. Stick with what the previous patch (I95f7b79f15) did and accept only
   inputs without a day of the week.
2. Add an option to the parser similar to the existing "monthNames"
   option and require callers to list all possible names. Only these
   will be accepted.
3. Make an entirely separate parser for Chinese dates.

Bug: T214002
Change-Id: Idab77dc1deba2fc5048647e7e4fcd1b92e6d9247