Make DateFormatParser accept more Asian/Chinese date formats
MediaWiki does have quite a lot of date formats that end with "(D)"
or "(l)" – the day of the week. These formats are currently entirely
rejected.

But the day of the week is not even needed to identify a date.

With this trivial change the parser acts as if the date formats do
not contain the day of the week. As a result input strings like
"2014年09月01日" are accepted, but "2014年09月01日 (一)" is not.

Bug: T214002
Change-Id: I95f7b79f155a9be0cc9fcb8db4498ba5994d45ea