Validate entity ID part of URIs in data values
The calendar of a time value, the globe of a globe coordinate value, and
the unit of a quantity value are expected to be item URIs (on Wikidata
in the former two cases, and on the local wiki in the latter case). Add
the missing validation for this. The new EntityUriValidator can also be
used to validate other entity types, if needed.

The previous implementation, which only validated the prefix of the URI
(up to /entity/), was using a message (wikibase-validator-bad-prefix)
which was not defined; define this message now, with the same first
parameter as before (the invalid value) in case anyone relies on it, but
also a second parameter (the prefix) for convenience. For the other
error conditions, we reuse existing messages.

Bug: T102840
Change-Id: I85cf40035dac8dbf749f7d6083ad85d2544a6f13