Don’t load Tainted References during QUnit tests
Apparently this causes problems during the QUnit tests, because they
want all code to be loaded before starting to run any of it, but here
we’re lazily loading additional ResourceLoader modules. I don’t think we
have any QUnit tests for tainted refs, so it’s probably fine to skip
loading the module in that case.

Bug: T330293
Change-Id: I44ff7c2bdd10680b71d453025d4a449cec5cc49e