Fix PHP 8 warning in Special:EntityData
When loaded without a subpage, it would cause a PHP warning from passing
null into preg_match(): EntityDataRequestHandler::handleRequest() and
EntityDataUriManager::parseDocName() are not supposed to be called with
null. This would happen, in particular, when using the new termbox,
which loads URLs looking like

> /w/index.php?title=Special:EntityData&id=Q2852&revision=6569&format=json

rather than

> /wiki/Special:EntityData/Q2852.json?revision=6569

I’m not changing any of the tests to pass null into the subpage, because
doing so somehow causes bizarre test failures in Wikibase and even in
some MediaWiki core tests (see T332129 for an excerpt and a link to full
CI outputs).

Change-Id: Ic1b780f12495ea3e37969e50fb156e8a57be183b