tests: Disable main cache in UsageTrackingIntegrationTest
Core enables most caches as HashBagOStuff when running PHPUnit,
and resets these between each test. WANObjectCache, the main entry
point for caching, is normally a wrapper around the "main" cache,
as configured by wgMainCacheType. However until now, core enabled
wgMainCacheType as HashBagOStuff, but then overrode WANObjectCache as
wrapping EmptyBagOStuff instead of Hash.

This is fixed in core change I53781a8c06ebb2583, and exposes one
test failure in Wikibase:

 1) Wikibase\Client\Tests\Integration\Usage\UsageTrackingIntegrationTest::testUpdateUsage
 Failed asserting that two arrays are equal.
 Array (
 -    0 => 'Q33#C.P123'
 +    0 => 'Q22#C.P123'
 )

This is likely result of something incorrectly sharing state, missing
a purge, or using a cache where it should bypass it. It's unclear
to me whether this is a test-only bug or a genuine bug. For now,
modify the test to behave the same as before by explicitly disabling
this cache.

Change-Id: I35cce61dc3ee90dcee3dd6f0b36f84133be029ed