bridge: fix type inference in test
TypeScript interpreted the return type of that inject call to be
`Getters<S>` instead of the actual type of the first parameter:
`StatementGetters`. By telling it that the statements are of type
StatementState, it manages to correctly infer the return type of the
inject call.

Also, we can use that opportunity to replace some more `any` types in
this file.

Change-Id: I991874bea963bd6d1858a3fd6221c9bced465867