Add the new selector for Kartographer maps
We need to tweak the selector for Kartographer because
data-mw='interface' conflicts with Parsoid-generated data-mw attributes.
This patch updates the selector for Wikibase (we intend to harmonize the
attribute between the legacy parser and Parsoid.)

The current version keeps both versions of the selector to account for
the fact that this patch and the Kartographer patch won't be merged at
the same time (this one needs to be first - will create the commit
message accordingly for the Kartographer patch), so we need to support
both versions for a little while.

However, there doesn't seem to be parsing cache considerations here: the
generated HTML is only cached in a small LRU cache, so there shouldn't
be any issue with cache expiration timings.

Bug: T331652
Change-Id: I5d52cddb3d672de0130f7a39b25f3fbbf09ac250