Move sitelinks collapse toggle into heading using placeholder
By default, MediaWiki core’s makeCollapsible module inserts a toggle
button as the first child of the collapsible element, which means it
becomes a sibling of the heading:

    div.wikibase-sitelinkgroupview.mw-collapsible
      span.mw-collapsible-toggle
      div.wikibase-sitelinkgroupview-heading-section
        ...
      div.mw-collapsible-content
        div.wikibase-sitelinklistview
          ul
          ...

This means that the toggle gets covered by the heading and also isn’t
sticky like the heading is. Moving the toggle inside the heading (still
as the first child) solves both of these issues, without any downsides
that I’ve noticed. (The previous change I911a49e345, commit fe10a2b825,
is therefore no longer necessary, and reverted here.)

This is implemented by adding a placeholder element to the DOM marking
where the toggle should go, in conjunction with a MediaWiki core change
(see Depends-On) that makes jquery.makeCollapsible take this placeholder
into account. MediaWiki core is still responsible for creating the real
link, with the translated and auto-updating text contents and all.

Bug: T329167
Change-Id: I51574be3efd8aa7c858e0ca0f27d2213d2f3cbb4
Depends-On: Ic2c222f139e56b4bd8c12490646d213fd9d25337