InjectRCRecordsJob: Remove transaction
This was originally a call to commitAndWaitForReplication() added in
change If195a35e9c (commit 24d45d6d0c), back when recentchanges records
were still injected as part of ChangeHandler. They were later moved to a
separate job in change I9100c17453 (commit 2987fe5bcf), but the commit
was kept; then, in change I58724c89df (commit 3aa96d61e2), this was
replaced with startAtomic() / endAtomic(), because the waiting for
replication was considered unnecessary. But as far as I understand, the
transaction isn’t necessary either, and since it sometimes causes
deadlocks in CI, let’s try removing it.

Bug: T299077
Change-Id: I17460e2fc926ef91f50242e5d3c090b395915ec8