REST: Run phpcs separately from the rest of WB
This adds a separate phpcs composer script for the rest-api/ directory,
and makes the REST API specific config an extension of the base
.phpcs.xml. That way the two will not interfere with each other, which
caused confusion in the past. See e.g.
I3abcef900e62732cef898a6d96203d4266efd671 and the discussion in comments
on I62d43dd96702544b67c62af5536d688ce9aaec49.

This does come with some drawbacks: phpcs needs to be invoked twice
(once for the base rules and once using the REST API specific ones),
and convenience composer scripts like phpcs-committed and phpcs-modified
don't work for the rest-api/ directory. Maybe we'll find a more elegant
solution in the future but for now this is still better than accidentally
interfering with code style rules outside of our directory.

Change-Id: I2d7b88b48cace5ef3865de26276a2ad19932334c