EntityUsageTable: Use UnionQueryBuilder
And pass around SelectQueryBuilder instances instead of SQL strings. In
the non-MySQL method, we can also use fetchRow() instead of iterating
over the result set manually. (fetchField() would also be an option.)

Bump the MediaWiki requirement, since UnionQueryBuilder is a brand new
class in core. (Bump it for both repo and client, since it doesn’t make
sense for the two to be different.)

Bug: T333690
Change-Id: I0376384eab0eba2b85156b297ce2e8d456dd0891
Depends-On: I14093100153c91baadf76c41b0dd4b896e77d222