tests: Add missing setName() in SitesModuleTest
Follows-up Wikibase 0ce52558c62d5 (Ic62e98fd590c33400)
and MediaWiki 2c93649672b (I2b2905d7452bd2).

See also ResourceLoader::getModule() the test suite of any Module
subclass in core. There are a number of things such as the name
that are required to be injected, or else things blow up. Some of
these one can afford to skip in certain lucky cases such as Config
and HookContainer are only used by certain methods that perhaps your
test doesn't call.

I'm adding this call here so that the core commit can be reverted
as a missing name is an error condition that should not be silently
accepted in production code. I'm open to a change that would e.g.
have getName() throw if not set if preferred.

Change-Id: I1cc5c54db42ca3a9da38ab56cf88cd71dc5b8343