Avoid array to string conversion in DataValueDeserializer
Check if the given value is a string before interpolating it into the
message; if it isn’t, then the value is even more invalid than usual,
and just put its type into the message. This fixes an “array to string
conversion” notice in production.

There’s already a TODO in the service wiring that “this should perhaps
be factored out into a class”, and this makes the case for that even
stronger IMHO (it’s strange to test this in the service wiring tests),
but let’s leave that for another change and fix the notice first.

Bug: T334719
Change-Id: I92493702d8931c643f7a92c85dd82d7da4c4612a