Show entity labels in parsed edit summaries in API requests
Aside from no longer ignore API requests per se, this does now stop
rendering label for Special:BadTitle. This **should** make no difference
in practice (aside from not reintroducing T111346), but if it does, then
we want the temporary feature config to cover _both_ changes. That is
the reason for the strange double-method in HtmlPageLinkRendererEndHookHandler.

Bug: T327062
Change-Id: I949863576e7efa6e68586f9159b14929d4d047d4