Add regression test for parser cache pollution (T111346)
We're prone to reintroduce a bug that pollutes browser cache with labels
in a specific language, which was bug T111346. Add a browser test that
checks for that.
It is a browser test, so that we make as little assumptions as possible
on how the API endpoint `stashedit` and parser cache work internally.

Bug: T327062
Change-Id: I40c5bffa9d36fd8995601ff592f124411d7ed34d