REST: Remove `Presentation` namespace
We recently got rid of our "presenter" classes in favor of just calling
the serializers in the route handlers. This leaves
`ErrorResponseToHttpStatus` as the only class in the namespace, which
in turn is only used by `ResponseFactory` in the `RouteHandlers`
namespace. Maintaining and documenting "presentation" as a whole layer
for this one class (which doesn't really do much) seems unnecessary, so
let's get rid of the namespace/directory/layer for now until we have a
real purpose for it.

Change-Id: I2daf57a717c493708dcbd9c46a0ff78538ce9e79