REST: Avoid collision with itself on repeated PUT
With this change, label and description validation is split into two
parts: input validation and item validation. The latter includes two
steps: a check that label and description aren't the same and a check
that no Item with the same label/description pair exists. Item
validation will be skipped, if the Item to be changed doesn't exist
or is a redirect, if the Item doesn't have both a label and a
description or if they remain unchanged. This way we avoid detecting
a 'collision' with the already existing item.

Bug: T335578
Change-Id: Ia2e09e9b98523a58680a7bbf4b0ba2aa32c48a63