Bump eslint-config-wikimedia from 0.24.0 to 0.25.0
This should allow most ES6 syntax.

A bunch of rules that now start complaining are skipped or disabled. The
only one I feel unsure about is the RegExp in templates.js, which looks
like it might be potentially problematic indeed (though I’m not sure how
user-controlled the input is); everything else is a false positive IMHO.

Change-Id: I6aa51d023bac4c2e5240aa7c1f176213801ac6d8