Drop RL targets from repo (except experts)
The code here is extracted from Iabd4aed6. Getting the service container
statically is not great, but we can clean that up separately in T336246.

The targets infrastructure for experts will be done in a separate commit.

The module 'wikibase.ui.entityViewInit' already is limited to desktop
already in this change, but not has its target dropped yet. The reason is
that there is effectively a false positive from the core
ResourcesTest::testUnsatisfiableDependencies test, due to the expert
modules still having there desktop target. So this will haven in a
follow-up change.

This for now also excludes removing the target from tainted-ref and
entitysearch, both of which deserve to be handled individually.

Bug: T324991
Change-Id: I90f81c02a01228d907aa9ef19b22da00753e54f4