Drop RL targets from Wikidata-Bridge modules
The construct of

```php
[
 'targets' => $clientSettings->getSetting( 'dataBridgeEnabled' ) ?
  [ 'desktop', 'mobile' ] :
  [],
]
```

was used to make sure that these modules are only available in
ResourceLoader if and only if the respective setting is enabled. (Which
is currently exclusively the case on cawiki.)

That functionality is now implemented more explicitly by checking for
the setting in the ResourceLoaderRegisterModules hook before adding the
modules.

This was also the only place that used the targets-option of
`MediaWikiConfigModule`.

Bug: T324991
Change-Id: Ia909dfe0144a05fb7823afb3a5e186e69344014c