Skip entity search module only on Minerva (and Vector-2022)
This slightly changes the behavior in edge cases, but for the better:

1. Minerva on desktop (either selected in the preferences or via ?useskin=)
   will no longer receive the 'wikibase.ui.entitysearch' module. Which
   is good, because that was unused anyway and just dead traffic.
2. Legacy Vector, Monobook and Timeless on m.wikidata.org, will receive
   the 'wikibase.ui.entitysearch' module, which means that their search
   will actually work again. However, afaik the only way to make use of
   them on m.wikidata.org is via ?useskin=, otherwise Minerva is forced.

Previously, this would essentially load based on "is m.wikidata.org or
not?". The separate search implementation for Minerva was added in
context of T85368.

Extracted from Iabd4aed6 where the main point was to drop the "targets"
declaration.

Bug: T324991
Change-Id: Ifc25e496d26371991717c832656ac1db723c5ad9