Run PHPStan only if core dependencies installed
This is a slightly naive (but effective!) fix for PHPStan failing in
environments where MediaWiki core or its dependencies are missing.
Alternatively we could remove PHPStan from `composer test` and ensure
that it only runs in jobs that include core and its depenencies, but
that would require more work and changes outside of Wikibase.git.

Skipping them silently does come at the risk of us not noticing if they
ever stop running in CI, but we would realize it eventually. Since the
tests detect structural and not functional issues, discovering them late
would only be annoying, but not critical.

Bug: T333454
Change-Id: Ic044fb02faefc2175babb74ea6fd230910d66712