REST: Rename "ItemData" to "ItemParts" to avoid naming collision
The `ItemData` ReadModel class was created to store optionally partial data of an Item. The `EntityRevisionLookupItemDataRetriever` is a class that implements *several* interfaces all related to the retrieval of item data and not just the `ItemDataRetriever` interface. The two different uses of "ItemData" is ambiguous and confusing for developers.

This patch renames "ItemData" to "ItemParts" for all relevant classes/interface (`ItemData`, `ItemDataBuilder`, `ItemDataRetriever`, `ItemDataSerializer`). Renaming `EntityRevisionLookupItemDataRetriever` was considered as an alternative solution, but no better name could be found. The new term "ItemParts" conveys the same meaning as "ItemData" did without the naming collision.

Change-Id: I5fd11cfb3b1f5faf9c55486630607c7c96440bec