Move jquery.wikibase.wbtooltip and dependencies to Lib
This module was previously in View, i.e. only available to Repo, but
change I02e075fb35 (commit 8fd203eb26) incorrectly started using it in
Client as well, breaking client-only wikis. Since that change can’t
easily be reverted (jquery.tipsy is already gone from MediaWiki core),
fix it for now by moving the module to Lib, making it available to both
Repo and Client. In the longer term, we should find a better solution
for this, as we really don’t want to have more ResourceLoader modules in
Lib than absolutely necessary.

Bug: T337081
Change-Id: I254949628b489381582c60efb9083ef15eda9f9b