Drop all Ruby browser tests, sadly effectively abandonware
Though these tests were a great step forward when created, nowadays (and
for years) Wikibase is the only repo to still use Ruby not JavaScript to
run its browser test suite, and despite the task being open to port them
nothing has happened at all for six years. It's time to accept that they
are not going to be ported, and that clearly the value to the team isn't
high enough to justify waiting. Instead, let's drop them for now, and so
reduce the CI runtime for developers, and complexity of maintaining this
part of the CI infrastructure. If the Wikibase team (or any other) later
wish to revive these tests in JavaScript, I and others would be happy to
assist.

[Note: If we're going to do this, we would first need to disable runs of
the legacy CI job which executes the Ruby browser tests.]

Bug: T180228
Bug: T220035
Change-Id: I56185e3e086c5a6e822e4e7caccec0647f98842f