REST: Create AssertItemExists for edit use cases
Read use cases, such as `GetItem`, call `GetLatestItemRevisionMetadata::execute()` which returns the revision metadata or throws an exception if the item can't be found or is a redirect. Edit use cases only need to check for item existence as new revision metadata will be created, when updating the item, and used instead. This patch adds the better named `AssertItemExists` for these edit use cases.

Change-Id: Ib8127d8377c4cc959b87077d93ec64de01b81c79