Use SelectQueryBuilder in DatabaseTermsCollisionDetector
This replaces the old DatabaseEntityTermsTableProvider with a custom
SelectQueryBuilder subclass, which does the necessary JOINs and also has
two methods to add conditions for terms.

I initially tried to just have the provider return a regular
SelectQueryBuilder with the JOINs already added, but then passing the
$entityIdColumn became awkward; with a dedicated subclass, we can have a
getEntityIdColumn() method on there, as well as move the “where” methods
from the collision detector to the SQB, where they’re better testable.

Bug: T312510
Change-Id: Ieee1ed498e0fdab917bd68317c15bff6d96731e5