Use SelectQueryBuilder in DatabaseTermInLangIdsResolver
I think I postponed this one a few times, but it’s not as difficult as
it looks: the callers in resolveTermsViaJoin() only pass in a single
table to JOIN, not an arbitrary list of JOINs, so it can be nicely
converted to ->join() after all.

As a bonus, resolveGroupedTermInLangIds() and resolveTermsViaJoin() now
pass their own method names into the query, instead of both showing up
as selectTermsViaJoin().

Bug: T312510
Change-Id: I42dcb2c0756c0d6b19862f8dccef156f1de88f2a