Remove repo prefixes from LinkFormatter test
Entity IDs can be formatted as HTML links on wikis other than the one
they're defined on. In such cases they're intended to look similar to
interwiki links. The interwiki link syntax (e.g. `[[wikidata:Q42]]`)
looks similar to the format of prefixed "foreign" entity IDs, but
they're different concepts.

The issue with ItemPropertyIdHtmlLinkFormatterTest was that it tested
this behavior with prefixed entity IDs, when in reality they're just
regular entity IDs that have a non-local "title".

Bug: T291823
Change-Id: Ifbb9ac58dcf55fee7c2b6476bb20becd99112721