SECURITY: Run edit filters on submit (undo/restore)
SubmitEntityAction, which implements the “undo” and “restore” actions,
doesn’t use the EditEntity interface (because SubmitEntityAction works
with redirects but EditEntity doesn’t support them), so we need to call
the EditFilterHookRunner ourselves to interact with e.g. AbuseFilter.

The EditFilterHookRunner interface is extended to also allow passing in
an EntityContent, so we don’t waste time (and complicate the code) by
having SubmitEntityAction extract an EntityDocument or EntityRedirect
from the EntityContent only for the hook runner to construct the content
again. Note that StatsdTimeRecordingEditFilterHookRunner didn’t even
properly support EntityRedirect yet (which doesn’t have a getType()
method; the only reason this didn’t blow up before is that the
StatsdTimeRecordingEditFilterHookRunner is only created and used by
MediawikiEditEntityFactory, but the EditFilterHookRunner service just
uses MediawikiEditFilterHookRunner directly), so fix that at the same
time to avoid problems later.

In the tests, to set up an edit filter that runs against all edits but
only blocks the one test case we add here, use the “Oslo” item handle
that doesn’t seem to have been used in any other tests as far as I can
tell. To use it with “undo” functionality, that test item now needs a
second revision added in ActionTestCase::makeTestItemData().

Bug: T250720
Change-Id: Id0260598b41c5c4683a446f478ac3dd84608b8fd